app_backend/requirements.txt
fastapi
uvicorn[standard]
joblib
tldextract
python-whois
dnspython
sentence-transformers
torch
xgboost
shap
requests
